% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibify_pkgs.R
\name{bibify_pkgs}
\alias{bibify_pkgs}
\title{Generate a bibliography of packages}
\usage{
bibify_pkgs(pkgs, csl, keep_intermediate = FALSE, ...)
}
\arguments{
\item{pkgs}{A character vector of packages}

\item{csl}{A csl file used to format the output bibliography.  Good place to
look for these is \url{https://github.com/citation-style-language/styles}}

\item{keep_intermediate}{A logical indicating if you would like to keep the
intermediate \code{.bib} and \code{.md} files.  Most useful if you would
like to get an output \code{.bib} or need to troubleshoot.  Default is
FALSE.}

\item{...}{Used to pass arguments to \link{rmarkdown::render}.  Most useful
for changing output format (e.g. \code{output_format = "word_document"}) or
output file name.}
}
\description{
This function creates an output bibliography from an input character vector
of package names.  The primary use case for this function is when working
with other authors and you need to provide a reference list of packages used.
}
\examples{
download.file("https://raw.githubusercontent.com/citation-style-language/styles/master/plos.csl",
                destfile = "plos.csl")
bibify_pkgs(c("dplyr","readr"),"plos.csl", output_format = "word_document",
            output_file = "my_package_bib.docx")
}
