% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byter.R
\name{byter}
\alias{byter}
\title{Byte Converter}
\usage{
byter(obj, units = c("b", "Kb", "Mb", "Gb", "B", "KB", "MB", "GB", "auto"),
  num = FALSE, ...)
}
\arguments{
\item{obj}{object to return size for.  Can also be a numeric (assumes bytes).}

\item{units}{units to convert to.}

\item{num}{boolean to switch converting an actual value as opposed to `object
size` object (the default)}
}
\description{
Simple wrapper to object size and format.
}
