% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_interest.R
\name{comp_interest}
\alias{comp_interest}
\title{Compound Interest}
\usage{
comp_interest(principal, apr, years, compound, recurr = 0)
}
\arguments{
\item{principal}{starting ammound}

\item{apr}{annual percentage rate (as a decimal)}

\item{years}{number of years to compound}

\item{compound}{frequency at which interest is compounded}

\item{recurr}{recurring payments made at same frequency as compounding}
}
\description{
Formula to calculate compound interest with or without recurring payments
}
