% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_poly.R
\name{create_poly}
\alias{create_poly}
\title{Interactively Create Polygon from active graphics device}
\usage{
create_poly(p4s = NULL)
}
\arguments{
\item{p4s}{proj.4 string for the new poly.  Should be taken from data showing in current graphics device}
}
\description{
Interactively Create Polygon from active graphics device
}
