% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_size.R
\name{dir_size}
\alias{dir_size}
\title{Return tibble with directory size}
\usage{
dir_size(path = ".", ...)
}
\arguments{
\item{path}{path to return directory size}
}
\description{
This function calculates total size for a specified directory.
}
\examples{
dir_size()                     
}
