% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_size.R
\name{dir_sizes}
\alias{dir_sizes}
\title{Return tibble with directory sizes}
\usage{
dir_sizes(paths, ...)
}
\arguments{
\item{paths}{paths for which to return directory sizes}
}
\description{
This function calculates total sizes for a vector of directories.
}
\examples{
dir_size()      
}
