% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_shp}
\alias{download_shp}
\title{function to download all available shapefile files from a URL}
\usage{
download_shp(shape_url, layer, outfile = layer)
}
\description{
function to download all available shapefile files from a URL
}

