% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_packages.R
\name{find_packages}
\alias{find_packages}
\title{Find all packages mentioned in a file
 
looks for all istances of "libray(package_name)" or "package_name::function".
Returns a unique listing of the mentioned packages.}
\usage{
find_packages(file)
}
\arguments{
\item{file}{a file to look for packages in}
}
\description{
Find all packages mentioned in a file
 
looks for all istances of "libray(package_name)" or "package_name::function".
Returns a unique listing of the mentioned packages.
}
\examples{
find_packages("lessons/01_rstudio.Rmd")
}
