% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_packages_file.R
\name{find_packages_folder}
\alias{find_packages_folder}
\title{Find called packages in a folder}
\usage{
find_packages_folder(folder, packages_r = TRUE)
}
\description{
This function searches a folder recursively for packages called by `library()`,
`require()`, or `::` that are within R, Rmd, or qmd files.  It returns a 
vector of all unique package calls and optionally writes out a "packages.R" file.
}
