% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_nlcd.R
\name{get_nlcd}
\alias{get_nlcd}
\title{Get a NLCD from USGS LANDFIRE program}
\usage{
get_nlcd(bbx = NULL, p4s = NULL)
}
\arguments{
\item{bbx}{a bounding box from \code{sp}}

\item{p4s}{a proj4string of projection to request image in.}
}
\value{
raster with NLCD colortable
}
\description{
Uses the USGS LANDFIRE NLCD/USGS_EDC_LandCover_NLCD (ImageServer) REST API.  Not sure
which year this is returning...
}
\examples{
\dontrun{
data(lake,package="quickmapr")
x_nlcd<-get_nlcd(bbox(buffer),proj4string(buffer))
}
#
}
\keyword{internal}

