% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_rstudiolog}
\alias{get_rstudiolog}
\title{Get RStudio Mirror logs}
\usage{
get_rstudiolog(date_range = c(Sys.Date() - 2, Sys.Date() - 1), folder = ".",
  missing = TRUE)
}
\arguments{
\item{date_range}{}

\item{folder}{}

\item{missing}{NOT YET IMPLEMENTED}
}
\description{
This function downloads the RStudio CRAN mirror logs for a date range to a
specified folder.  Returns a data frame for all downloaded files
}
\examples{
x<-get_rstudiolog(c(Sys.Date()-5,Sys.Date()-1),folder="rstudio_logs",missing=TRUE)
}

