% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_oops.R
\name{git_oops}
\alias{git_oops}
\title{Git oops...}
\usage{
git_oops(big_file, keep_file = TRUE, ignore_file = TRUE)
}
\description{
This function uses the git filter-branch method of removing an accidentally 
commited BIG file from the git history so that you may get back to happily
commiting your little files!  It assumes you want to keep the big file, but 
no longer track it.  As such it does not remove the file and adds it to 
.gitignore.
}
