% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cv_compare.R
\name{holdout}
\alias{holdout}
\title{Some tests of different cv techniques...}
\usage{
holdout(x, y, perc, R = 1)
}
\description{
Some tests of different cv techniques...
}
\examples{
x<-rnorm(1000,10,2)
y<-x+rnorm(1000,0,2)
holdout(x,y,0.75,1000)
loo(x,y)
Holdout
}

