% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mirror_repo.R
\name{one_drive_mirror}
\alias{one_drive_mirror}
\title{Set up path for a OneDrive remote}
\usage{
one_drive_mirror(repo_name = basename(getwd()), jeff = TRUE)
}
\description{
Defaults on this are for jeff
}
\keyword{internal}
