% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_perform_io.R
\name{sys_perform_io}
\alias{sys_perform_io}
\title{Speed test for read and write}
\usage{
sys_perform_io(n = 10, delete_file = TRUE)
}
\arguments{
\item{n}{Number of times to repeat the read/write operations}

\item{delete_file}{Delete the random csv generated for this test.}
}
\description{
This function records I/O times for the current working directory
}
\examples{
sys_perform_io(5)                     
}
