% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{Counting Words in R}
\usage{
word_count(txtfile, skip_lines = NULL)
}
\arguments{
\item{txtfile}{path and filename of a file to count the words}

\item{skip_lines}{a numeric vector indicating lines of the file to skip.
                  These lines are removed prior to removing YAML and code chunks.
                  Default is NULL.}
}
\description{
Function to count words in an Rmd file.  YAML headers and Code Chunks are ignored and
optional lines to skip can be included.
}
\examples{
my_file <- system.file('extdata/test.Rmd',package='miscPackage')
word_count(my_file)
}

