% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_shp.R
\name{download_shp}
\alias{download_shp}
\title{function to download all available shapefile files from a URL}
\usage{
download_shp(shape_url, layer, outfile = layer)
}
\description{
function to download all available shapefile files from a URL
}

