% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_shp.R
\name{download_shp}
\alias{download_shp}
\title{function to download all available shapefile files from a URL}
\usage{
download_shp(shape_url, layer, outfolder = ".")
}
\description{
function to download all available shapefile files from a URL
}
\examples{
\dontrun{
download_shp("ftp://ftp.granit.sr.unh.edu/pub/GRANIT_Data/Vector_Data/Administrative_and_Political_Boundaries/d-nhsenatedists/2012","NHSenateDists2012")
download_shp("http://jwhollister.com/iale_open_science/files","LocPt")
}
}

