% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sizes.R
\name{file_sizes}
\alias{file_sizes}
\title{Return tibble with directory size}
\usage{
file_sizes(path = ".", ...)
}
\arguments{
\item{path}{path to return directory size}
}
\description{
This function returns file sizes for all files in a directory
}
\examples{
dir_size()                     
}
