% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rstudiolog.R
\name{get_rstudiolog}
\alias{get_rstudiolog}
\title{Get RStudio Mirror logs}
\usage{
get_rstudiolog(
  date_range = c(Sys.Date() - 2, Sys.Date() - 1),
  folder = ".",
  missing = TRUE
)
}
\arguments{
\item{missing}{NOT YET IMPLEMENTED}
}
\description{
This function downloads the RStudio CRAN mirror logs for a date range to a 
specified folder.  Returns a data frame for all downloaded files
}
\examples{
\dontrun{
x<-get_rstudiolog(c(Sys.Date()-30,Sys.Date()-1),folder='rstudio_logs',missing=TRUE)
}
}
