% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_coords.R
\name{remove_coords}
\alias{remove_coords}
\title{Simplifies a polygon by removing every n vertices}
\usage{
remove_coords(sp_obj, every)
}
\arguments{
\item{sp_obj}{}

\item{every}{}
}
\description{
Simplifies a polygon by removing every n vertices
}

