% Generated by roxygen2 (4.0.2): do not edit by hand
\name{update_pubs}
\alias{update_pubs}
\title{Read Google Scholar Profile and save citation counts to a file}
\usage{
update_pubs(id, scholar_csv)
}
\description{
This will read in google scholar Profile.  If the file already exists (i.e. this
has been run before), then a new column is added with the number of new citations
}
\examples{
\dontrun{
update_pubs("Fn9BjfIAAAAJ","~/projects/google_scholar.csv")}
}

