% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/word_count.R
\name{word_count}
\alias{word_count}
\title{Counting Words in R}
\usage{
word_count(txtfile)
}
\arguments{
\item{txtfile}{path and filename of a file to count the words}
}
\description{
Function to count words in a text file or a character vector.
}
\examples{
my_file <- system.file('extdata/test.Rmd',package='miscPackage')
word_count(my_file)
}

