% Generated by roxygen2 (4.0.2): do not edit by hand
\name{word_count}
\alias{word_count}
\title{Counting Words in R}
\usage{
word_count(txt, file = FALSE)
}
\arguments{
\item{txt}{character that either contains either a vector to count or path
and filename of a file to count the words}

\item{file}{boolean to indicate if txt is a file or not (probably better
way to implement, but this was quick).}
}
\description{
Function to count words in a text file or a character vector.
}

