% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_lines.R, R/word_count.R
\name{word_count}
\alias{word_count}
\title{Break lines of an input file}
\usage{
word_count(txtfile, skip_lines = NULL)

word_count(txtfile, skip_lines = NULL)
}
\arguments{
\item{txtfile}{path and filename of a file to count the words}

\item{skip_lines}{a numeric vector indicating lines of the file to skip.
These lines are removed prior to removing YAML and code chunks.
Default is NULL.}

\item{outfile}{Output file to save}

\item{width}{integer indicating width at which to set line break.
Defaults to 80 characters.}

\item{yaml}{Logical indicating whether or not YAML should be included.
Not yet implemented.}

\item{chunks}{Logical indicating whether or not code chunks should be
included.  Not yet implemented.}
}
\description{
Function to break lines at a specified width.  Finds nearest white space to 
80 characters and breaks the line there.

Function to count words in an Rmd file.  YAML headers and Code Chunks are ignored and 
optional lines to skip can be included.
}
\examples{
my_file <- system.file('extdata/test.Rmd',package='miscPackage')
word_count(my_file)
#@export
my_file <- system.file('extdata/test.Rmd',package='miscPackage')
word_count(my_file)
}
